feat #46: PR findings: update comments, remove unneeded double conversion, simplify mapping to GenericTimedDataPoint

+++ app/src/main/java/org/htwk/pacing/backend/predictor/preprocessing/GenericTimedDataPoint.kt
@@ -4,6 +4,14 @@ import kotlinx.datetime.Instant
 import org.htwk.pacing.backend.database.DistanceEntry
 import org.htwk.pacing.backend.database.HeartRateEntry
 
+/**
+ * A generic representation of a data point that has a specific time and a numeric value.
+ * This class serves as a standardized format for various types of vital data,
+ * such as heart rate or distance, allowing them to be processed uniformly.
+ *
+ * @property time The [Instant] at which the data point was recorded.
+ * @property value The numeric value of the data point, represented as a [Double].
+ */
 data class GenericTimedDataPoint(
     val time: Instant,
     val value: Double
@@ -15,6 +23,6 @@ data class GenericTimedDataPoint(
 
     constructor(src: DistanceEntry) : this(
         time = src.end,
-        value = src.length.inMeters().toDouble()
-    ) // example: magnitude
+        value = src.length.inMeters()
+    )
 }
